Refactor: (main) 폴더 삭제
